refactor : school API 구조 변경
