+++ src/main/java/com/zerobase/school/school/repository/SchoolRepository.java
@@ -1,7 +1,6 @@
-package com.zerobase.school.user.repository;
+package com.zerobase.school.school.repository;
 
-import com.zerobase.school.user.domain.entity.School;
-import org.springframework.data.domain.Example;
+import com.zerobase.school.school.domain.School;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
